Use int instead of float

+++ apps/openmw/mwrender/renderingmanager.cpp
@@ -299,7 +299,7 @@ namespace MWRender
         globalDefines["radialFog"] = Settings::Manager::getBool("radial fog", "Shaders") ? "1" : "0";
         globalDefines["groundcoverAnimation"] = Settings::Manager::getBool("animation", "Groundcover") ? "1" : "0";
 
-        float groundcoverDistance = (Constants::CellSizeInUnits * Settings::Manager::getFloat("distance", "Groundcover") - 1024) * 0.93;
+        float groundcoverDistance = (Constants::CellSizeInUnits * Settings::Manager::getInt("distance", "Groundcover") - 1024) * 0.93;
         globalDefines["groundcoverFadeStart"] = std::to_string(groundcoverDistance * Settings::Manager::getFloat("fade start", "Groundcover"));
         globalDefines["groundcoverFadeEnd"] = std::to_string(groundcoverDistance);
 
@@ -1334,7 +1334,7 @@ namespace MWRender
 
         if (mGroundcoverWorld)
         {
-            float groundcoverDistance = Constants::CellSizeInUnits * Settings::Manager::getFloat("distance", "Groundcover");
+            int groundcoverDistance = Constants::CellSizeInUnits * Settings::Manager::getInt("distance", "Groundcover");
             mGroundcoverWorld->setViewDistance(groundcoverDistance * (distanceMult ? 1.f/distanceMult : 1.f));
         }
     }
